int-float

+++ lesson_4/exercise_1.py
@@ -3,8 +3,8 @@ digital_2 = 42.99
 digital_3 = -0.6
 digital_4 = 0.4
 digital_5 = 5.5
-print(round(digital_1))
-print(round(digital_2))
-print(round(digital_3))
-print(round(digital_4))
-print(round(digital_5))
+print(float(round(digital_1)))
+print(float(round(digital_2)))
+print(float(round(digital_3)))
+print(float(round(digital_4)))
+print(float(round(digital_5)))
